// src/api/generated/models/apiUserStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */

export type ApiUserStatus = typeof ApiUserStatus[keyof typeof ApiUserStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApiUserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
} as const;